move SignUp component to Home directory
