+++ src/pages/index.tsx
@@ -7,7 +7,7 @@ import { HowtoPage } from './Howto/Howto'
 import { HomePage } from './Home/Home'
 import { NotFoundPage } from './NotFound/NotFound'
 import { DiscussionsPage } from './Discussions'
-import { SignUpPage } from './SignUp'
+import { SignUpPage } from './Home/SignUp'
 import ScrollToTop from './../components/ScrollToTop/ScrollToTop'
 import { EventsPage } from './Events/Events'
 import Header from './common/Header/Header'
